refact(#33): url 관련 내용 util로 분리
